+++ tests/MetricExecutionTime.cpp
@@ -212,20 +212,19 @@ bool MetricExecutionTime::v_Test(istream &pStdout, istream &pStderr)
         success = false;
     }
 
-    // Average the found execution times.
-    double avgTime = 0.0;
+    // Find the minimum of the execution times.
+    double minTime = times[0];
     for (unsigned int i = 0; i < times.size(); ++i)
     {
-        avgTime += times[i];
+        minTime = (times[i] < minTime) ? times[i] : minTime;
     }
-    avgTime /= times.size();
 
     // If a skip flag is set (due to no matching hostname) then return
     // successful test and output the average time.
     if (m_match.m_skip)
     {
-        cerr << "Average execution time for host "
-             << boost::asio::ip::host_name() << ": " << avgTime << endl;
+        cerr << "Minimum execution time for host "
+             << boost::asio::ip::host_name() << ": " << minTime << endl;
         return success;
     }
 
@@ -233,15 +232,15 @@ bool MetricExecutionTime::v_Test(istream &pStdout, istream &pStderr)
              "No test conditions defined for execution time.");
 
     // Check that the average time is within the tolerance.
-    if (fabs(avgTime - boost::lexical_cast<double>(m_match.m_value)) >
+    if (fabs(minTime - boost::lexical_cast<double>(m_match.m_value)) >
             m_match.m_tolerance &&
         !m_match.m_skip)
     {
         cerr << endl;
         cerr << "Failed tolerance match." << endl;
-        cerr << "  Expected avg execution time: " << m_match.m_value << " +/- "
+        cerr << "  Expected min execution time: " << m_match.m_value << " +/- "
              << m_match.m_tolerance << endl;
-        cerr << "  Actual avg: " << avgTime << endl;
+        cerr << "  Actual min: " << minTime << endl;
 
         for (unsigned int i = 0; i < times.size(); ++i)
         {
@@ -316,14 +315,13 @@ void MetricExecutionTime::v_Generate(istream &pStdout, istream &pStderr)
     if (matched)
     {
         // Average the found execution times and set it as the accepted value.
-        double avgTime = 0.0;
+        double minTime = sampleTimes[0];
         for (unsigned int i = 0; i < sampleTimes.size(); ++i)
         {
-            avgTime += sampleTimes[i];
+            minTime = (sampleTimes[i] < minTime ? sampleTimes[i] : minTime);
         }
-        avgTime /= sampleTimes.size();
 
-        okValue.m_value = to_string(avgTime);
+        okValue.m_value = to_string(minTime);
         m_match         = okValue;
     }
     else